requirements.txt
python-telegram-bot==22.0
dotenv==0.9.9
anthropic==0.49.0
pinecone==6.0.2
telegramify-markdown==0.5.1